Add files via upload
I document step by step process of Exploratory data analysis for medical data.  The purpose is to estimate the annual medical costs for new insurance customers. We make estimation based on the information of historical insurance customers. 

In the next notebook, I will develop linear models to make the estimate.